bundler: Share Loader / Resolver between multiple testing rules
